move recursive bytes type encoding test to signed_data_test.go
